# Align `TextBuilder` characters by their baselines during addition/removal