--- a/osu.Framework/Text/TextBuilderGlyph.cs
+++ b/osu.Framework/Text/TextBuilderGlyph.cs
@@ -16,6 +16,7 @@ namespace osu.Framework.Text
         public readonly float XOffset => ((fixedWidth - Glyph.Width) / 2 ?? Glyph.XOffset) * textSize;
         public readonly float YOffset => Glyph.YOffset * textSize;
         public readonly float XAdvance => (fixedWidth ?? Glyph.XAdvance) * textSize;
+        public readonly float Baseline => Glyph.Baseline * textSize;
         public readonly float Width => Glyph.Width * textSize;
         public readonly float Height => Glyph.Height * textSize;
         public readonly char Character => Glyph.Character;