… osu.Framework/Text/TextBuilder.cs