Previously characters were *implicitly aligned* using their `YOffset`s, and that works when only one font is used, but not when more than one font gets used inside one `SpriteText`/`TextBuilder`.

This PR aims to resolve the "multiple fonts" case by offsetting their `DrawRectangle`s based on the difference between the `Baseline` of the new character and the previous characters.

I've worked on this initially so that font-mixed sprite texts with #3271 don't look badly misaligned (as I noticed after implementing #3271), but apparently, there are already alignment improvements osu!-side with this PR alone (although slightly unnoticable):

- Before:
![image](https://user-images.githubusercontent.com/22781491/133373547-b409f35c-27e6-455a-af43-cf990156951d.png)

- After:
![image](https://user-images.githubusercontent.com/22781491/133373555-8c905a4e-11b0-45fd-99c8-8e0d9217c55b.png)

(notice `(TV Size)` aligned correctly with the other glyphs)

I have added in-depth test coverage in `TextBuilderTest` which had the perfect setup for having so, with all possible scenarios I can think of.

Along with test coverage, I've also added a benchmark including the regular "no baseline adjustment" cases, and "one baseline adjustment every 10 characters" case, and the results seem reasonable.

|                                    Method |     Mean |    Error |   StdDev |
|------------------------------------------ |---------:|---------:|---------:|
|                             AddCharacters | 19.99 us | 0.280 us | 0.323 us |
|       AddCharactersWithDifferentBaselines | 31.02 us | 0.219 us | 0.183 us |
|                       RemoveLastCharacter | 24.90 us | 0.493 us | 0.506 us |
| RemoveLastCharacterWithDifferentBaselines | 38.14 us | 0.570 us | 0.505 us |

I can see this being further optimised by offseting new characters to negative `X/Y`s rather than offsetting all the other previous characters in the line, but that might be a bit complicated to deal with in the `SpriteText_DrawNode` (especially given that `TextBuilder` can have multiple lines), therefore I've went with the simplest way of for looping all characters in the line for now.