@peppy 2021-09-16
this is missing a breaking change clause for the updated method signature, will add.